requirements.txt
streamlit==1.20.0
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.0.123
cassio==0.1.0
cassandra-driver==3.28.0
openai==1.3.0
anthropic==0.3.0
sentence-transformers==2.2.2
tiktoken==0.5.1